fix: filter issues solved
