+++ frontend/src/contexts/app-context-interface.ts
@@ -1,12 +1,10 @@
-import { ongoingStatusKeys } from '@components/filtering/errand-filter';
 import { Asset } from '@interfaces/asset';
-import { ErrandsData, IErrand } from '@interfaces/errand';
+import { emptyErrandList, ErrandsData, IErrand } from '@interfaces/errand';
 import { UiPhase } from '@interfaces/errand-phase';
-import { ErrandStatus } from '@interfaces/errand-status';
+import { ErrandStatus, getStatusKeys, ongoingStatuses } from '@interfaces/errand-status';
 import { MessageNode } from '@interfaces/message';
 import { Notification as CaseDataNotification } from '@interfaces/notification';
 import { User } from '@interfaces/user';
-import { emptyErrandList } from '@services/casedata-errand-service';
 import { Admin, emptyUser } from '@services/user-service';
 import { createContext } from 'react';
 
@@ -88,7 +86,7 @@ export const AppContext = createContext<AppContextInterface>({
   municipalityId: '',
   setMunicipalityId: () => {},
 
-  selectedErrandStatuses: ongoingStatusKeys as ErrandStatus[],
+  selectedErrandStatuses: getStatusKeys(ongoingStatuses) as ErrandStatus[],
   setSelectedErrandStatuses: () => {},
 
   notifications: [],
